requirements.txt
# GACI — Python Dependencies

# Anthropic Claude API
anthropic>=0.25.0

# Web framework — webhook receiver
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# HTTP client — Jira API calls
httpx>=0.27.0

# Vector store — knowledge repository indexing
chromadb>=0.4.24

# YAML frontmatter parsing and writing
python-frontmatter>=1.1.0
PyYAML>=6.0.1

# Environment variable loading
python-dotenv>=1.0.0

# Jira API client
jira>=3.6.0

# Markdown processing
markdown>=3.6

# Date/time handling
python-dateutil>=2.9.0

# Audit log (SQLite ORM)
sqlalchemy>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # also used for test client
